V5 rc1

+++ oonipipeline/tests/conftest.py
@@ -23,6 +23,7 @@ from oonipipeline.db.connections import ClickhouseConnection
 from oonipipeline.db.create_tables import make_create_queries
 from oonipipeline.fingerprintdb import FingerprintDB
 from oonipipeline.netinfo import NetinfoDB
+from oonipipeline.settings import config
 
 from ._fixtures import SAMPLE_MEASUREMENTS
 
@@ -77,6 +78,7 @@ def temporal_dev_server(request):
 
 @pytest.fixture
 def datadir():
+    config.data_dir = str(DATA_DIR)
     return DATA_DIR
 
 
@@ -157,5 +159,10 @@ def db_notruncate(clickhouse_server):
 def db(clickhouse_server):
     db = create_db_for_fixture(clickhouse_server)
     for _, table_name in make_create_queries():
+        # Ignore the fingerprints_dns table, since it's a remote table
+        if table_name == "fingerprints_dns":
+            continue
         db.execute(f"TRUNCATE TABLE {table_name};")
+
+    config.clickhouse_url = db.clickhouse_url
     yield db
